RF: place BinaryFxFeatureMeasure tests into a separate tests file
